Replace Spring Retry use to core retry
This commit replaces the use of Spring Retry by the core retry support
in Spring Framework 7. This is a breaking change mostly in configuration
that is detailed below.

Some of the features used from Spring Retry have no equivalent in Spring
Framework, and a tailored solution for Spring AMQP has been implemented
in this commit. This leads to the change in the following hook points:

* `AbstractAdaptableMessageListener` can be configured by the newly
introduced `MessageRecoveryCallback` instead of Spring Retry's
`MethodInvocationRecoverer`. The new callback directly provides the
Message, replyTo Address, and Exception.
* The `RecoveryCallback` in `RabbitTemplate` is replaced by an interface
with the same name.
* `AbstractRetryOperationsInterceptorFactoryBean` can be configured with
a `RetryPolicy`, rather than a `RetryTemplate`.

Spring AMQP uses stateless and stateful retry interceptors, with
corresponding features in Spring Retry. These are replaced by:

* `StatelessRetryOperationsInterceptor` that invokes recovery if
provided.
* `StatefulRetryOperationsInterceptor` provides a similar interceptor
as the Spring Retry equivalent, except that it is configured with
Spring AMQP-specify callbacks.

These interceptors are configured with a `RetryPolicy`, rather than
a `RetryTemplate`. The new `RetryTemplate` in Spring Framework is
operating on a `RetryPolicy` can can't be further configured. This
has an impact on `RetryInterceptorBuilder` where a custom `BackOff`
can no longer be specified. Instead, a consumer of the
RetryPolicy.Builder is provided that offers configuration for it and
more.

Using dedicated callback also shows inconsistencies with the Null-safety
support. The most annoying example is MessageKeyGenerator that should
be nullable despite its contract stating that it should not. Spring
AMQP has actually a code path to deal with stateful retries when the
key is null.

A major difference in default is that Spring Retry has a default
policy of 3 attempts: the initial invocation and 2 retries. Spring
Framework has a default RetryPolicy of 3 attempts, that does not
include the initial invocation. With 4 attempts in total, there is an
off-by-one change in various areas, mostly in tests, to configure a
retry policy with 2 attempts.